refactor: Tried to add miniSearch
