+++ package.json
@@ -40,6 +40,7 @@
     "markdown-it-emoji": "^2.0.0",
     "markdown-it-footnote": "^3.0.2",
     "mdi-material-ui": "^6.21.0",
+    "minisearch": "^3.0.2",
     "moment": "^2.29.1",
     "nanoid": "^3.1.20",
     "noty": "^3.2.0-beta",
