+++ src/components/DeleteNoteDialog.tsx
@@ -10,7 +10,7 @@ import { TabNode } from "flexlayout-react";
 import React from "react";
 import { useTranslation } from "react-i18next";
 import { CrossnoteContainer } from "../containers/crossnote";
-import { Note } from "../lib/notebook";
+import { Note } from "../lib/note";
 
 interface Props {
   open: boolean;
